# [MRG+1] Improve strip_accents_unicode performance when strings do not need stripping