@jnothman 2016-05-22
This seems sensible to me. +1