@gglanzani 2016-05-22
@agramfort @jnothman Thanks for looking at it!