@jnothman 2016-05-22
Thank you for the contribution.

On 23 May 2016 at 05:37, Giovanni Lanzani <notifications@github.com> wrote:

> @agramfort <https://github.com/agramfort> @jnothman
> <https://github.com/jnothman> Thanks for looking at it!
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub
> <https://github.com/scikit-learn/scikit-learn/pull/6805#issuecomment-220851539>
>